requirements.txt
streamlit
Pillow
rembg
numpy
opencv-python
scikit-image
